# Much faster prompt processing for k-quants (ARM_NEON)